+++ Race Element.Data/Games/rFactor2/DataMapper/LocalCarMapper.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RaceElement.Data.Games.rFactor2;
+namespace RaceElement.Data.Games.rFactor2.DataMapper;
 
 internal static class LocalCarMapper
 {
